Update patch set 11
Patch Set 11:

Yes, What I mean is that we can just uses a global variable for the table, then we can use a constructor function to initialize the global variable once. Then for each process, it will be initialized once for a progress. But definitely, if the tests are testing many progress instances, for each progress, it needs to do the table initialization.

Patch-set: 11
Reviewer: Gerrit User 1011275 <1011275@d5d70762-12d0-45a1-890d-524b12d3f735>