Update patch set 6
Patch Set 6:

(1 comment)

Regarding the test, we've talked offline on this. Unfortunately we are not sure if there is way to reproduce it every time within functional tests. It requires the IO not to finish in time for closing. Lots of iops to large nvme device, helped us to showcase it most of the time on dev setup. Not sure there is 100% reproducible way to achieve it in functional tests.

I'm fine with proceeding without the test.

Patch-set: 6
Reviewer: Gerrit User 1011221 <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>