bdevperf: Provide unregister callback when opening target bdev.
Bdevperf up to now didn't use unregister callback when opening target
block deice. This is unsafe if test scenario involves target bdev
unregistering voluntarily.

To be on the safe side, this patch registers a callback to drain I/O
queue and shutdown target (and to end the run if that's the last live
target) in case the target unregisters during runtime.

Change-Id: I0dad2f4a7cf4acb22b363e23e4b670ca73d8c6f2
Signed-off-by: Andrey Kuzmin <akuzmin@jetstreamsoft.com>