Update patch set 2
Patch Set 2: Code-Review-1

Hi Darek, so I added this for two reasons (a) because this module is supposed to be an example and (b) I had one test case that didn't work without it, I can't remember what it was because it was so long ago but it took me forever to debug it until Ben explained to me how some of this buffer stuff works (this was a long time ago).

I'd prefer to keep it in and maybe just add more content about when/why its needed vs not?

Patch-set: 2
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1