Bdev/QoS: add the user controllable unmap limit
In the case where unmap or similar operations (discard/trim/write_zeroes)
received, it also needs to be limited.

Added a new and user controllable limit type for the unmap operation.
The unmap limit is separate from the write limit. In order to avoid the
timeout from the larger unmap operation, it is expecting to set a high
limit for the unmap bandwidth limit.

Change-Id: Ifffd159bfe3129e5d975c63f0ae43c7f4d48de3b
Signed-off-by: GangCao <gang.cao@intel.com>