Added possibility to set spdt_mempool private data
Added wrapper for dpdk_mempool_create which allows setting mempool private
data size, and getter to retrive private pointer data.

OCF requires memory retrived from mempool to be filled with zeros.

When mempool is created, we save element size to its private data to use it
later in memeset when new memory is retrived.

This is fix for issue #671

Signed-off-by: Michal Mielewczyk <michal.mielewczyk@intel.com>
Change-Id: If09d35b2f3eaaf0827ac82ac74476cdfb0baa331