reduce: pass buf/iov to _issue_backing_ops
When a chunk cannot be compressed enough to save one I/O
unit, we will just write the uncompressed data to disk.  In
this case, we want the helper routine to use the decomp_buf
instead of the comp_buf.  Instead of passing flags to specify
which to use, have the callers pass the actual buf and iov array
from the req that should be used for the operation.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I1537130577943e19fe4dfcee58eac453eca3c29b