Update patch set 7
Patch Set 7:

Hi Chunyang,

Thanks for the update.
I should have said before submitting the patch but how about the following?

It looks that this RPC outputs information like union, so introducing "fields", "details", or "breakdown" may be helpful to understand.

This is an idea and please feel free to keep the current patch. But please update doc/jsonrpc.md even so.

{
  "jsonrpc": "2.0",
  "id": 1,
  "result":  {
    "version": "19.04-pre",
    "fields": {
      "major": "19",
      "minor": "4",
      "suffix": "-pre"
    }
  }
}

Patch-set: 7