perf: associate each namespace to each core.
In some cases, we could configure perf with more than 1 cores and
currently we used the round robin algorithm to assoicate the cores
and namespaces. In this case, one namespace may only be accessed
through one core.

This patch adds a new parameter to let all cores to access all
 the namespaces.

Change-Id: I731f683dc30541257ea169dc5190189101eab894
Signed-off-by: Liang Yan <liang.z.yan@intel.com>