vhost: support transferring inflight buffer between qemu and backend
When guest system is running, if spdk crashes or killed
by signals, access to the block dev may hung.

Because IO processing is out of order, We can't determine
which IOs are actually completed and which IOs are in
inflight. In other words, the order of IO recorded in avail
and used arrays may be different.This may lead to the
resubmission of some completed IOs, while other inflight IOs
are not submitted, resulting in IO hung.

To support this, This patch introduces two new messages
VHOST_USER_GET_INFLIGHT_FD and VHOST_USER_SET_INFLIGHT_FD
to support transferring a shared buffer between qemu and backend.
so when spdk crashes, the memory still exists. At the
begining of the initiation of spdk, it will resubmit unfinished
requests at first, and then go on with the normal poller work.

Change-Id: Ie83e16cd2dd5e61d636b9487bdaaea9f0d97d57b
Signed-off-by: lilin24 <lilin24@baidu.com>
Signed-off-by: Ni Xun <nixun@baidu.com>
Signed-off-by: Zhang Yu <zhangyu31@baidu.com>