app/trace: optimize by multiple merge sorting
Trace files can be generated directly by spdk app, or
by spdk_trace_record to poll trace shm.
Both of them should be well ordered from the start of
the trace buffer to the end, or from the middle to the
end and then from the start back to the middle.
Original sorting on trace entries is a full scale sorting,
which is not necessary and occupied too much memory if
the size of trace file is huge.

Change-Id: I2b4b833cf434c949aed03d46b299d5890853bf82
Signed-off-by: Xiaodong Liu <xiaodong.liu@intel.com>