bdev/aio : avoid spdk assert when bdev subsystem is initialized failed.
This happens when we start using the configuration file with [Nvme]' target
to start vhost.

We assume that the order of the bdev subsystem to initialize
the bdev modules is as follows: nvme, aio. If the nvme fails to initialize,
vhost will call the all bdev modules include uninitialized to clean up the
resource function.

In bdev_aio_fini, we use the 'spdk_io_device_unregister' to clean up the resource
directly without judging that this module is calling the 'bdev_aio_initialize'

To support this, we introduce a global variable to determine if the aio module is
initialized.

Change-Id: I6034331524ca29bef765e8381f576996b8434a13
Signed-off-by: Tianyu Yang <yangtianyu2@huawei.com>