bdev: Clean up module resources that failed to initialize
Currently, if the Bdev subsystem initialization fails, the initialization failed
module does not call the Fini function cleanup resource.

However, there are modules such as aio,null that use the 'spdk_io_device_register'
function when initializing, and if these modules do not call the fini function,
it will cause resource leak when exiting the app.

To avoid this, introduce a global variable that records the failed module
and ensure that module calls the fini function.

Change-Id: I582f3ff511a51764ec74371dd952e23989f0568a
Signed-off-by: yangtianyu 00383654 <yangtianyu2@huawei.com>