Update patch set 35
Patch Set 35:

That should be all of the comments.  Wrt backing dev IO size, 4K works fine now so that was something else early on in the dev of the module that was wrong.  Setting any kind of length in dst mbufs doesn't seem to be done in the cdev perf or test code so I removed calculating it and send Fiona an email to confirm

Patch-set: 35
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>