bdev: defer unregister untill *all* descriptors are closed
We used to wait only for those descriptors which
specified the hotremove notification callback. The
bdev could've been removed before the descriptor
was closed and the subsequent spdk_bdev_close would
simply segfault. We modify spdk_bdev_unregister to
always wait for all descriptors to be closed before
actually unregistering the bdev. This consolidates
the bdev unregister behavior and will be inevitable
after the upcoming refactor.

Change-Id: I9b358209c6abd301b6fe8660e27bc6fa4ef485d6
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>