bdev: allow closing descriptors from any thread
Bdev descriptors could be closed only from the same thread
that opened them. Not anymore.

We move the responsibility of closing all descriptors and
finalizing the bdev unregister to a single _remove_desc()
function. The same function is meant to be called for
notifying a descriptor about hotremoval, handle closing
it, and also proceeding with the deferred bdev unregister.
This function can be called only on the same thread that
opened the descriptor, but the trick is it can be simply
scheduled from any other thread. It utilizes the existing
desc state flags to know what to do, and those can be set
from any thread holding the bdev's lock.

Change-Id: I49fde296ba691124c7b2065332b4d9cfbbfd3d4c
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>