[WIP] test/iscsi_tgt: fix UBSAN error in VPP
This patch is NOT intended to be merged in this shape.

This issue is already solved in VPP and requires cherry-picking commit
f5dc9fbf814865b31b52b20f5bf959e9ff818b25 in VPP.
(see
https://review.gerrithub.io/c/spdk/spdk/+/394174/105/test/common/config/vm_setup.sh#222)

This patch temporaly turns off UBSAN for now, while CI is not fully updated yet.

Change-Id: I3bbc264bbc8d9747e81d70c1c6f74d6c88e20ae1
Signed-off-by: Tomasz Kulasek <tomaszx.kulasek@intel.com>