bdev/compress: add UT for _compress_operation
Add coverage for _compress_operation() as well as
general test setup required for this one, additional
mocks, etc. Some of what we do in the mocked DPDK
functions make look like we're just testing our own
test code when we assert values, but in order to
confirm values like chained mbufs and total pkt len
we have to add some basic functionality in the mocked
functions.

Change-Id: Ief08f9b71ee7a836f6026d26517f5faa5f9d51ce
Signed-off-by: paul luse <paul.e.luse@intel.com>