lib/thread: Fix wrong posix thread name by SPDK thread scheduler
Before enabling SPDK thread scheduler,
each posix thread name was reactor_% (% was the core number).

After enabling SPDK thread scheduler, the name of the master posix
thread is reactor_% (% is the max core number), and the name of the
slave posix threads are lcore-slave-% (% is the current core number).

SPDK threads are light-weight threads - sometimes also called
green threads or fibers, and so are independent from posix
threads.

So it will be reasonable if SPDK don't change posix thread name,
and the name of the master thread is the name of the SPDK application
(e.g. nvmf_tgt) and the name of the slave threads are the name
added by DPDK (e.g. lcore-slave-%).

Hence this patch removes _set_thread_name() in lib/thread/thread.c.

Change-Id: I26e8dabc73e163c9f74e18b3640cf54954603b1f
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>