Update patch set 9
Patch Set 9: Code-Review-1

I need to know the sequence of steps that cause the rqpair to not have a port/device at this point in time. This function is processing a request, so the rqpair must already have a port/device set up for that to happen. If this is happening on the shut down path, then I think we have some clean up happening out of order that we should fix instead.

Patch-set: 9
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1